If published date is not set then send null for this property to server (#1409)
* feat: allow published date to be null in generate-version-json and send-version-to-epp

In `generate-version-json.ts` and `send-version-to-epp.ts` we now allow the published date to be null.

* feat: handle null values in generate-version-json

In the generate-version-json.ts file, added null coalescing operator to handle possible null values for 'publishedDate' and 'publishedYear'. This change ensures that the function will not throw an error when these fields are null.

* feat: update EPP API docker image in docker-compose.yaml

Update the EPP API docker image from 'ghcr.io/elifesciences/enhanced-preprints:master-416bf3bb-20230912.1944' to 'ghcr.io/elifesciences/enhanced-preprints-server:master-001111cd-20231004.1022' in docker-compose.yaml.