feat: add approval query to importDocmaps workflow (#2465)
* feat: add approval query to importDocmaps workflow

This commit introduces a new query 'awaitingApproval' to the importDocmaps
workflow. The approval state is now initially undefined and will be set to
null when the docMapThreshold is exceeded. The query handler is set to return
true when the approval state is null, indicating that approval is being awaited.

* feat: update query handling in import-docmaps workflow

This commit modifies the 'importDocmaps' function in the
'import-docmaps.ts' file. It replaces the 'approvalQuery' with a
'thresholdQuery' and adjusts the corresponding handler. The handler for
'thresholdQuery' now returns a message indicating how many docMapIdHashes
are awaiting approval if the number exceeds the defined threshold.

* feat: add threshold query response type in import-docmaps.ts

This commit introduces a new type 'ThresholdQueryResponse' to the
import-docmaps.ts file. It also modifies the 'thresholdQuery' to use this
new type. The handler for 'thresholdQuery' has been updated to return an
object with 'awaitingApproval' count and 'docMapUrls' array, instead of
just the count.

* feat: Refactor importDocmaps function in import-docmaps.ts

- Initialize 'approval' variable to null
- Streamline the conditional logic in setHandler function for thresholdQuery
- Remove redundant 'approval' reset after checking docMapThreshold condition

* feat: refactor thresholdMet in import-docmaps.ts

A new function `thresholdMet` is introduced in `import-docmaps.ts` to check if the length of `docMapIdHashes` array is more than the `docMapThreshold`. This function is used to refactor the existing condition checks which were previously scattered across multiple places in the file. This should improve readability and maintainability.

* style: apply correct spacing to import statement in import-docmaps.ts