Update docker-compose and transformXML configuration [elifesciences/enhanced-preprints-issues#1159]
Update the docker-compose file to use a new version of `temporalio/admin-tools` and `enhanced-preprints-encoda` images. Also, modify the `transformXML` function in `convert-xml-to-json.ts` file to include an optional header for passthrough based on the `xsltTransformPassthrough` configuration. Additionally, make changes in `config.ts` file to include the `xsltTransformPassthrough` configuration and update `encodaDefaultVersion` to '1.0.6'.