Add type for author notes and meta in the article (#4445)
* add type for author notes and meta in the article

* Update docker-compose and transformXML configuration [elifesciences/enhanced-preprints-issues#1159]

Update the docker-compose file to use a new version of `temporalio/admin-tools` and `enhanced-preprints-encoda` images. Also, modify the `transformXML` function in `convert-xml-to-json.ts` file to include an optional header for passthrough based on the `xsltTransformPassthrough` configuration. Additionally, make changes in `config.ts` file to include the `xsltTransformPassthrough` configuration and update `encodaDefaultVersion` to '1.0.6'.

* Remove fixme comment from xsltTransformPassthrough in config.ts [elifesciences/enhanced-preprints-issues#1159]

* Make label optional and add meta to content structure

Update 'generate-version-json.ts' to include 'meta' from 'contentStruct'. In 'send-version-to-epp.ts', make 'label' in 'authorNotes' optional.

* Correct the array syntax in 'send-version-to-epp' module [elifesciences/enhanced-preprints-issues#1159]

* Add empty headers to mockedAxios.post in convert-xml-to-json tests [elifesciences/enhanced-preprints-issues#1159]

* Undo unnecessary changes [elifesciences/enhanced-preprints-issues#1159]

---------

Co-authored-by: Will Byrne <will.alex.byrne@gmail.com>